Initial commit.
The beginnings of a support of the (2019 version of the) IEEE-754
standard for floating points.

The standard is (for the moment) implemented as an interface, or two
interfaces to be more accurate (corresponding to the required and
optional part of the standard, respectively), and the actual
implementation is implemented as an implentation of that interface.

In the future, I will likely ditch that approach, as this generality
doesn't seem to add any practical value (who wants to implement his own
IEEE floats anyway?).